+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-## [2.11.2] - 2020/06/30
+## [2.12.1] - 2020/06/30
 
-* ReduxAction.status and ReduxAction.hasFinished. (Search for "Action status" in README.md) 
+* MockStore (still experimental, in `mock_store.dart`), lets you mock or disable actions/reducers during tests. 
+  (See section `Mocking actions and reducers` in README.md).
+* ReduxAction.status and ReduxAction.hasFinished. (Search for "Action status" in README.md).
 
 ## [2.11.1] - 2020/06/12